Changed mach_CycleTime to be an unsigned int to remove floating point from kernel.

+++ CVSROOT/kernel/mach/spur.md.save/mach.h
@@ -194,7 +194,7 @@ extern	int	mach_LastUserStackPage;
  * is intended for use by the devTimer module.
  */
 
-extern 	double	mach_CycleTime;
+extern 	unsigned int	mach_CycleTime;
 
 
 #endif _MACH
